backend/requirements.txt
Django==4.0.10
djangorestframework==3.14.0
django-cors-headers==3.14.0
